created a surveyor repository for storing list of surveyors

+++ src/main/java/sysc4806/group27/minisurveymonkey/repository/SurveyorRepository.java
@@ -3,5 +3,6 @@ package sysc4806.group27.minisurveymonkey.repository;
 import org.springframework.data.repository.CrudRepository;
 import sysc4806.group27.minisurveymonkey.model.Surveyor;
 
-public interface SurveyorRepository extends CrudRepository<Surveyor, Integer>  {
+
+public interface SurveyorRepository extends CrudRepository<Surveyor, Long>  {
 }
